+++ js/directives/lucca/user-picker.js
@@ -42,7 +42,7 @@
 
 	var userPickerTemplate = "<ui-select theme=\"bootstrap\"" +
 	"class=\"lui {{size}} \" on-select=\"onSelect()\" on-remove=\"onRemove()\" ng-disabled=\"controlDisabled\">" +
-	"<ui-select-match placeholder=\"{{ ::placeholder }}\" allow-clear=\"true\">" +
+	"<ui-select-match placeholder=\"{{ ::placeholder }}\" allow-clear=\"{{allowClear}}\">" +
 		"<div ng-if=\"!$select.selected.isAll\">{{ $select.selected.firstName }} {{$select.selected.lastName}}</div>" +
 		"<div ng-if=\"$select.selected.isAll\">{{ 'LUIDUSERPICKER_ALL' | translate }}</div>" +
 	"</ui-select-match>" +
@@ -51,7 +51,7 @@
 
 	var userPickerMultipleTemplate = "<ui-select multiple theme=\"bootstrap\" " +
 	"class=\"lui {{size}} input\" on-select=\"onSelect()\" on-remove=\"onRemove()\" ng-disabled=\"controlDisabled\" close-on-select=\"false\">" +
-	"<ui-select-match placeholder=\"{{ ::placeholder }}\" allow-clear=\"true\">{{$item.firstName}} {{$item.lastName}} " +
+	"<ui-select-match placeholder=\"{{ ::placeholder }}\" allow-clear=\"{{allowClear}}\">{{$item.firstName}} {{$item.lastName}} " +
 		"<small ng-if=\"$item.hasHomonyms && getProperty($item, property.name)\" ng-repeat=\"property in displayedProperties\"><b>{{property.label | translate}}</b> {{getProperty($item, property.name)}} </small>" +
 		"<small ng-if=\"$item.isFormerEmployee\" translate  translate-values=\"{dtContractEnd:user.dtContractEnd}\">LUIDUSERPICKER_FORMEREMPLOYEE</small>" +
 	"</ui-select-match>" +
@@ -73,6 +73,7 @@
 				onSelect: "&",
 				onRemove: "&",
 				controlDisabled: "=",
+				allowClear: "@",
 				/*** FORMER EMPLOYEES ***/
 				showFormerEmployees: "=", // boolean
 				/*** HOMONYMS ***/
@@ -104,6 +105,7 @@
 				upCtrl.asyncPagination = false;
 				upCtrl.useCustomFilter = !!attrs.customFilter;
 				upCtrl.displayCustomInfo = !!attrs.customInfo || !!attrs.customInfoAsync;
+				scope.allowClear = !!attrs.allowClear ? scope.allowClear : false;
 
 				scope.$watch(function() {
 					return (ngModelCtrl.$viewValue || {}).id;
@@ -140,6 +142,7 @@
 				onSelect: "&",
 				onRemove: "&",
 				controlDisabled: "=",
+				allowClear: "@",
 				/*** FORMER EMPLOYEES ***/
 				showFormerEmployees: "=", // boolean
 				/*** HOMONYMS ***/
@@ -169,6 +172,7 @@
 				upCtrl.asyncPagination = false;
 				upCtrl.useCustomFilter = !!attrs.customFilter;
 				upCtrl.displayCustomInfo = !!attrs.customInfo || !!attrs.customInfoAsync;
+				scope.allowClear = !!attrs.allowClear ? scope.allowClear : false;
 
 				scope.$watchCollection(function() {
 					return ngModelCtrl.$viewValue;